+++ changelog/adapter/ironsource-mobileads/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 5.9.0.0
+
+### Updated
+* Added support for Yandex Mobile Ads SDK 5.9.0
+
 ## Version 5.8.0.0
 
 ### Updated
@@ -9,7 +14,7 @@ All notable changes to this project will be documented in this file.
 ## Version 5.7.0.0
 
 ### Updated
-* Added support for Yandex Mobile Ads SDK 5.6.0
+* Added support for Yandex Mobile Ads SDK 5.7.0
 
 ## Version 5.6.0.0